requirements.txt
numpy
opencv-contrib-python
opencv-python
openpyxl
pandas
pillow
pyttsx3
reportlab